Make desktop feature optional for tests
