+++ src-tauri/Cargo.toml
@@ -15,15 +15,11 @@ tauri-build = { version = "1.0.4", features = [] }
 [dependencies]
 serde_json = "1.0"
 serde = { version = "1.0", features = ["derive"] }
-tauri = { version = "1.0.5", features = ["api-all"] }
+tauri = { version = "1.0.5", features = ["api-all"], optional = true }
 rhai-sci = { version = "0.1.9", features = ["smartcore", "nalgebra", "rand", "io"] }
 rhai-ml = "0.1.2"
 rhai = { version = "1.9.0", features= ["sync"] }
 
 [features]
-# by default Tauri runs in production mode
-# when `tauri dev` runs it is executed with `cargo run --no-default-features` if `devPath` is an URL
-default = [ "custom-protocol" ]
-# this feature is used used for production builds where `devPath` points to the filesystem
-# DO NOT remove this
-custom-protocol = [ "tauri/custom-protocol" ]
+default = []
+desktop = ["tauri", "tauri/custom-protocol"]
